Added cancellation token support
All rewritten methods now accept a CancellationToken (in the future we
can actually check whether it is used internally).

When calling external methods (i.e. not those rewritten by us), we check
whether an overload accepting a cancellation token exists and call it.
Otherwise we check and call the overload without the cancellation token.